Display contacts in chat list. 1. Added event listener to the Topic class (this allow propagating events up to the UI). 2. Modified the ContactsViewController to (a) subscribe to "me" topic and (b) display the fetched subscriptions.
This is an intermediate commit. More commits to come: topics need to be displayed properly.
Also, currently, UI is using a 1 second timer to fetch/display topics.
This needs to be fixed so that we have a listner in the Tinode class that triggers
a subsciption fetch right after we've subscribed to "me" topic.