Substitute tinodeHasRunBefore for tinodeAppMetaVersion.
Context:
This will allow us to force the user to logout when the app updates.
Recent commits 22234a017e76b59fd19fedd5bc544ec31b0754e3 and
9b4c84122c18192b9dd99b594ada751f02fe4e95 changed the location of
both the settings data and the keychain. Both moved to the tinodios application
group. Whenever the user updates the app, it will read the connection
settings from the new location - and thus end up with the default values, not
what the user set originally.
In order to mitigate this, force the user to re-login explicitly,
so he can see what server he's trying to connect to in case of auth errors.